decompose tuple
